demonstrate  merge conflict

+++ abc.cpp
@@ -1,7 +1,9 @@
+//Comment to demonstrate merge conflict
+
 #include<stdio.h>
 int main()
 {
-    short int i = 0;
+    int i = 0;
     for(i<=5 && i>=-1; ++i; i>0)
         printf("%u,", i);
